Fixes to make my changes work
Changed Player Base combat detection loop from fixed update to update to catch all button presses.

+++ Assets/C#/Entity/Combat/MeleeCombat.cs
@@ -10,7 +10,7 @@ public class MeleeCombat : MonoBehaviour
     public float timeSinceLastHit = 0;
     public float meleeOffset;
     public int staminaMax = 100;
-    public float stamina = 100;
+    public float stamina = 100f;
     public int staminaCost = 15;
     public float staminaRechargeRate = 2f;
 
@@ -18,7 +18,7 @@ public class MeleeCombat : MonoBehaviour
 
     public void Update()
     {
-        staminaIndicator = GetStamina();
+        staminaIndicator = (int)GetStamina();
     }
 
     public bool CanShoot()
@@ -26,7 +26,7 @@ public class MeleeCombat : MonoBehaviour
         return stamina >= staminaCost;
     }
     
-    public int GetStamina()
+    public float GetStamina()
     {
         return Mathf.Min(stamina + (float)((Time.time - timeSinceLastHit) * staminaRechargeRate), staminaMax);
     }
